+++ selfdrive/car/hyundai/interface.py
@@ -25,7 +25,7 @@ class CarInterface(CarInterfaceBase):
     ret.safetyModel = car.CarParams.SafetyModel.hyundai
     ret.radarOffCan = RADAR_START_ADDR not in fingerprint[1]
 
-    ret.openpilotLongitudinalControl = Params().get_bool("DisableRadar") and candidate in [CAR.SONATA, CAR.PALISADE]
+    ret.openpilotLongitudinalControl = Params().get_bool("DisableRadar") and candidate in [CAR.SONATA, CAR.SONATA_HYBRID, CAR.PALISADE, CAR.SANTA_FE]
     ret.safetyParam = 0
 
     # Most Hyundai car ports are community features for now
@@ -42,7 +42,7 @@ class CarInterface(CarInterfaceBase):
 
     ret.longitudinalTuning.kpV = [0.1]
     ret.longitudinalTuning.kiV = [0.0]
-    ret.stoppingDecelRate = 2.0
+    ret.stopAccel = 0.0
     ret.startAccel = 0.0
 
     ret.longitudinalActuatorDelayUpperBound = 1.0 # s