+++ selfdrive/controls/lib/lead_mpc_lib/lead_mpc.py
@@ -4,6 +4,7 @@ import math
 import numpy as np
 
 from common.realtime import sec_since_boot
+from common.numpy_fast import clip
 from selfdrive.swaglog import cloudlog
 from selfdrive.modeld.constants import T_IDXS
 from selfdrive.controls.lib.drive_helpers import MPC_COST_LONG, CONTROL_N
@@ -20,10 +21,10 @@ MPC_T = list(np.arange(0,1.,.2)) + list(np.arange(1.,10.6,.6))
 N = len(MPC_T) - 1
 
 
-def RW(v_ego, v_l):
+def desired_follow_distance(v_ego, v_lead):
   TR = 1.8
   G = 9.81
-  return (v_ego * TR - (v_l - v_ego) * TR + v_ego * v_ego / (2 * G) - v_l * v_l / (2 * G))
+  return (v_ego * TR - (v_lead - v_ego) * TR + v_ego * v_ego / (2 * G) - v_lead * v_lead / (2 * G)) + 4.0
 
 
 def gen_lead_model():
@@ -84,21 +85,16 @@ def gen_lead_mpc_solver():
   ocp.cost.yref_e = np.zeros((3, ))
 
   x_lead, v_lead = ocp.model.p[0], ocp.model.p[1]
-  G = 9.81
-  TR = 1.8
-  desired_dist = (v_ego * TR
-                  - (v_lead - v_ego) * TR
-                  + v_ego*v_ego/(2*G)
-                  - v_lead * v_lead / (2*G))
-  dist_err = (desired_dist + 4.0 - (x_lead - x_ego))/(sqrt(v_ego + 0.5) + 0.1)
+  desired_dist = desired_follow_distance(v_ego, v_lead)
+  dist_err = (desired_dist - (x_lead - x_ego))/(sqrt(v_ego + 0.5) + 0.1)
 
   # TODO hacky weights to keep behavior the same
   ocp.model.cost_y_expr = vertcat(exp(.3 * dist_err) - 1.,
-                                  ((x_lead - x_ego) - (desired_dist + 4.0)) / (0.05 * v_ego + 0.5),
+                                  ((x_lead - x_ego) - (desired_dist)) / (0.05 * v_ego + 0.5),
                                   a_ego * (.1 * v_ego + 1.0),
                                   j_ego * (.1 * v_ego + 1.0))
   ocp.model.cost_y_expr_e = vertcat(exp(.3 * dist_err) - 1.,
-                                  ((x_lead - x_ego) - (desired_dist + 4.0)) / (0.05 * v_ego + 0.5),
+                                  ((x_lead - x_ego) - (desired_dist)) / (0.05 * v_ego + 0.5),
                                   a_ego * (.1 * v_ego + 1.0))
   ocp.parameter_values = np.array([0., .0])
 
@@ -185,7 +181,7 @@ class LeadMpc():
       t += dt
 
   def init_with_sim(self, v_ego, lead_xv, a_lead_0):
-    a_ego = min(0.0, -(v_ego - lead_xv[0,1]) * (v_ego - lead_xv[0,1]) / (2.0 * lead_xv[0,0] + 0.01) + a_lead_0)
+    a_ego = min(0.0, -2 * (v_ego - lead_xv[0,1]) * (v_ego - lead_xv[0,1]) / (2.0 * lead_xv[0,0] + 0.01) + a_lead_0)
     dt =.2
     t = .0
     x_ego = 0.0
@@ -201,6 +197,7 @@ class LeadMpc():
       self.solver.set(i, 'x', np.array([x_ego, v_ego, a_ego]))
 
   def update(self, carstate, radarstate, v_cruise):
+    self.crashing = False
     v_ego = self.x0[1]
     if self.lead_id == 0:
       lead = radarstate.leadOne
@@ -210,7 +207,15 @@ class LeadMpc():
     if lead is not None and lead.status:
       x_lead = lead.dRel
       v_lead = max(0.0, lead.vLead)
-      a_lead = lead.aLeadK
+      a_lead = clip(lead.aLeadK, -5.0, 5.0)
+
+      # MPC will not converge if immidiate crash is expected
+      # Clip lead distance to what is still possible to brake for
+      MIN_ACCEL = -3.5
+      min_x_lead = ((v_ego + v_lead)/2) * (v_ego - v_lead) / (-MIN_ACCEL * 2)
+      if x_lead < min_x_lead:
+        x_lead = min_x_lead
+        self.crashing = True
 
       if (v_lead < 0.1 or -a_lead / 2.0 > v_lead):
         v_lead = 0.0
@@ -248,7 +253,7 @@ class LeadMpc():
     self.j_solution = np.interp(T_IDXS[:CONTROL_N], MPC_T[:-1], list(self.u_sol[:,0]))
 
     # Reset if goes through lead car
-    self.crashing = np.sum(self.lead_xv[:,0] - self.x_sol[:,0] < 0) > 0
+    self.crashing = self.crashing or np.sum(self.lead_xv[:,0] - self.x_sol[:,0] < 0) > 0
 
     t = sec_since_boot()
     if self.solution_status != 0: