+++ selfdrive/debug/uiview.py
@@ -4,7 +4,7 @@ import cereal.messaging as messaging
 from selfdrive.manager.process_config import managed_processes
 
 if __name__ == "__main__":
-  services = ['controlsState', 'deviceState', 'pandaState']  # the services needed to be spoofed to start ui offroad
+  services = ['controlsState', 'deviceState', 'pandaState', 'carParams']
   procs = ['camerad', 'ui', 'modeld', 'calibrationd']
 
   for p in procs:
@@ -15,6 +15,7 @@ if __name__ == "__main__":
   msgs = {s: messaging.new_message(s) for s in services}
   msgs['deviceState'].deviceState.started = True
   msgs['pandaState'].pandaState.ignitionLine = True
+  msgs['carParams'].carParams.openpilotLongitudinalControl = True
 
   try:
     while True: