+++ RELEASES.md
@@ -2,6 +2,8 @@ Version 0.8.10 (2021-XX-XX)
 ========================
  * Škoda Kamiq 2021 support thanks to jyoung8607!
  * Škoda Karoq 2019 support thanks to jyoung8607!
+ * Volkswagen Arteon 2021 support thanks to jyoung8607!
+ * Volkswagen California 2021 support thanks to jyoung8607!
 
 Version 0.8.9 (2021-09-14)
 ========================